dashboard/pyproject.toml
[tool.poetry]
name = "plantsitter_dashboard"
version = "0.1.0"
description = "A dashboard to take care of the planty bois"
authors = ["LSaldyt <lucassaldyt@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
dash = "^1.19.0"
dash-devices = "^0.1.3"
Quart = "^0.14.1"
quart-compress = "^0.2.1"
requests = "^2.25.1"
pandas = "^1.2.2"
websocket_client = "^0.57.0"
twilio = "^6.54.0"
oauth2 = "^1.9.0"
googleAPI = "^0.1.0"
google-api-python-client = "^2.0.2"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.3"
pymongo = "^3.11.3"
selenium = "^3.141.0"
wikipedia = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
